1_Single/requirements.txt
python-dotenv
camel-ai
